Fix for problem that we had in class 19-10-21
There was nothing major wrong with the code. The problem was with
the learning rate that we were using. I think it was 2 and this
was way too high. I have changed it to 0.002 in this code and only
drawn the line if the x2 values don't get out of hand